requirements.txt
scrapy==2.11.2
scrapy-playwright==0.0.40
playwright==1.44.0
playwright-stealth==1.0.6
python-dotenv==1.0.1
pyyaml==6.0.1
w3lib==2.1.2
unicodedata2==15.1.0
fake-useragent==1.5.1
itemadapter==0.9.0
